--- client/src/components/UserRecipe/UserRecipe.js
@@ -1,23 +1,0 @@
-import React from "react";
-
-const UserRecipe = () => {
-    const [name, setname] = useState("")
-    const[selectedFile, setSelectedFile]= useState(null);
-  return (
-    <div className="App">
-      <form>
-        <input type="text"
-         value={name}
-         onChange={(e) => setName (e.target.value)}
-          />
-
-        <input type="file"
-        value={selectedFile}
-        onChange={(e) => setSelectedFile(e.target.files[0])}
-        />
-      </form>
-    </div>
-  );
-};
-
-export default UserRecipe;
