+++ client/src/pages/Cookbook.js
@@ -1,10 +1,17 @@
 import React from "react";
 import RecipeCard from "../components/RecipeCard";
 import UserRecipe from "../components/UserRecipe";
-
+import SearchBar from "../components/SearchBar"
 
 function CookBook(){
-    return null;
-}
+    return( 
+    <>
+        <SearchBar
+        />
+        <UserRecipe
+        />
+    </>
+    )
+};
 
-export default CookBook
+export default CookBook;
